Helicopter lands at targeted position. Player can't call helicopter from the top of steep slopes anymore
A suited landing area was defined as an area where the landing area
collider did not hit any objects or terrain. This allowed the player to
call the helicopter when standing on the top of steep slopes. This has
been fixed by adding checkpoints within the collider and comparing their
relative height above the ground. If the height difference is too big,
the area is not suited for a helicopter landing.